fix errors

+++ src/item_group.cpp
@@ -209,8 +209,7 @@ void Item_modifier::modify( item &new_item ) const
     }
 
     new_item.set_damage( rng( damage.first, damage.second ) );
-    double random_dirt = static_cast<double>( rng( static_cast<int>( dirt.first ),
-                         static_cast<int>( dirt.second ) ) );
+    int random_dirt = rng( dirt.first, dirt.second );
     if( new_item.is_gun() && random_dirt > 0 ) {
         new_item.set_var( "dirt",  random_dirt );
         new_item.faults.emplace( "fault_gun_dirt" );
